pyproject.toml
[tool.poetry]
name = "systemmonitoringtool"
version = "0.1.0"
description = ""
authors = ["ConorBobbleHat <conorb@firstpartners.net>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.10.5"
mypy = "^1.14.1"
colorlog = "^6.9.0"

[tool.poetry.group.collector.dependencies]
psutil = "^6.1.1"
websockets = "^14.2"
aiohttp = "^3.11.12"
requests = "^2.32.3"
appdirs = "^1.4.4"

[tool.poetry.group.server.dependencies]
flask = "^3.1.0"
flask-sock = "^0.7.0"
flask-sqlalchemy = "^3.1.1"
django = "^5.1.5"
channels = {extras = ["daphne"], version = "^4.2.0"}
django-ninja = "^1.3.0"
django-ninja-crud = "^0.6.2"
django-cors-headers = "^4.7.0"

[tool.poetry.group.dev.dependencies]
types-requests = "^2.32.0.20241016"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
